// modules/Delusoire/sort-plus/settings.d.ts
export declare const CONFIG: Record<string, never> & {
    preventDuplicates: boolean;
} & {
    descending: boolean;
} & {
    artistAllDiscography: boolean;
} & {
    artistTopTracks: boolean;
} & {
    artistPopularReleases: boolean;
} & {
    artistSingles: boolean;
} & {
    artistAlbums: boolean;
} & {
    artistCompilations: boolean;
} & {
    artistLikedTracks: boolean;
} & {
    artistAppearsOn: boolean;
} & {
    lastFmUsername: string;
} & {
    LFMApiKey: string;
} & {
    sortedPlaylistsFolderUri: string;
};